Remove BlogSelectorViewController
