--- WordPress/Classes/ViewRelated/Blog/Blog Selector/BlogSelectorViewController.h
@@ -1,32 +1,0 @@
-#import <UIKit/UIKit.h>
-
-@import CoreData;
-
-NS_ASSUME_NONNULL_BEGIN
-
-typedef void (^BlogSelectorSuccessHandler)(NSManagedObjectID *selectedObjectID);
-typedef void (^BlogSelectorSuccessDotComHandler)(NSNumber *dotComId);
-typedef void (^BlogSelectorDismissHandler)(void);
-
-@interface BlogSelectorViewController : UITableViewController
-
-- (instancetype)initWithSelectedBlogObjectID:(nullable NSManagedObjectID *)objectID
-                              successHandler:(nullable BlogSelectorSuccessHandler)successHandler
-                              dismissHandler:(nullable BlogSelectorDismissHandler)dismissHandler;
-
-- (instancetype)initWithSelectedBlogDotComID:(nullable NSNumber *)dotComID
-                              successHandler:(BlogSelectorSuccessDotComHandler)successHandler
-                              dismissHandler:(nullable BlogSelectorDismissHandler)dismissHandler;
-
-@property (nonatomic, assign) BOOL displaysOnlyDefaultAccountSites;
-@property (nonatomic, assign) BOOL displaysNavigationBarWhenSearching;
-@property (nonatomic, assign) BOOL displaysCancelButton;
-@property (nonatomic, assign) BOOL shouldHideSelfHostedSites;
-@property (nonatomic, assign) BOOL dismissOnCancellation;
-@property (nonatomic, assign) BOOL dismissOnCompletion;
-
-@property (nonatomic, copy) BlogSelectorSuccessHandler successHandler;
-
-@end
-
-NS_ASSUME_NONNULL_END
